backend/requirements.txt
# ============================================
# Web Framework
# ============================================
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0

# ============================================
# LLM + Agents
# ============================================
langchain==0.2.5
langchain-core==0.2.9
langgraph==0.1.14
langchain-google-genai==1.0.6
langchain-groq==0.1.6
langchain-community==0.2.5
langchain-chroma==0.1.1
langsmith==0.1.77

# ============================================
# Tools
# ============================================
tavily-python==0.3.3
e2b-code-interpreter==0.0.10

# ============================================
# Vector Store
# ============================================
chromadb==0.5.3

# ============================================
# Database (SQLite dev / PostgreSQL prod ready)
# ============================================
sqlmodel==0.0.19
aiosqlite==0.20.0
sqlalchemy==2.0.30

# ============================================
# Redis (Upstash serverless REST)
# ============================================
upstash-redis==1.1.0

# ============================================
# Auth & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
slowapi==0.1.9

# ============================================
# Validation & Config
# ============================================
pydantic==2.7.4
pydantic-settings==2.3.3
email-validator==2.2.0

# ============================================
# Logging
# ============================================
structlog==24.2.0

# ============================================
# HTTP Client
# ============================================
httpx==0.27.0

# ============================================
# Embeddings
# ============================================
google-generativeai==0.7.1
